Fix error implementation for KDKDK-4th mode, use tree_nb to calculate gradient correction for only EP particles (no SIMD yet); Add gradient correction for external force from galpy. Fix wrong data type for scale_density in gas_drag mode

+++ src/external_force.hpp
@@ -27,7 +27,7 @@ public:
     IOParams<double> polytropic_exponent; // gamma
 #ifdef GALPY
     IOParams<long long int> galpy_gaspot_index;
-    IOParams<long long int> scale_density;
+    IOParams<double> scale_density;
 #else
     IOParams<double> gas_density; 
     IOParams<double> decay_time;
